feat: add wayback viewer timeline scrubber
